cambio de los daos de biginteger a integer afectaba automaticamente a los model por lo cual tambien fueron adaptados

+++ src/java/Empleate/logica/CompanyModel.java
@@ -44,7 +44,7 @@ public class CompanyModel {
     }
     public Company findCompanyByID(int id)throws Exception{
         try{
-        return compDAO.findById(BigInteger.valueOf(id));
+        return compDAO.findById(id);
         }catch(Exception e){}
         return null;
     }
